requirements.txt
Django==1.9.7
django-bootstrap-form==3.3
djangorestframework==3.6.3
mysqlclient==1.3.9
redis==2.10.6
tornado==4.4.2
